项目功能扩展-微博登录图片验证码添加&绑定用户直接跳转首页&未绑定用户跳转绑定页面.

+++ front_end_pc/js/sina_callback.js
@@ -12,7 +12,8 @@ var vm = new Vue({
         error_phone_message: '',
         error_sms_code_message: '',
 
-
+        image_code_id: '', // 图片验证码id
+        image_code_url: '',
 
         sms_code_tip: '获取短信验证码',
         sending_flag: false, // 正在发送短信标志
@@ -24,7 +25,7 @@ var vm = new Vue({
         access_token: ''
     },
     mounted: function(){
-        // 从路径中获取qq重定向返回的code
+        // 从路径中获取微博重定向返回的code
         var code = this.get_query_string('code');
         axios.get(this.host + '/oauth/sina/user/?code=' + code, {
                 responseType: 'json',
@@ -63,7 +64,28 @@ var vm = new Vue({
             }
             return null;
         },
+        // 生成uuid
+        generate_uuid: function(){
+            var d = new Date().getTime();
+            if(window.performance && typeof window.performance.now === "function"){
+                d += performance.now(); //use high-precision timer if available
+            }
+            var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function(c) {
+                var r = (d + Math.random()*16)%16 | 0;
+                d = Math.floor(d/16);
+                return (c =='x' ? r : (r&0x3|0x8)).toString(16);
+            });
+            return uuid;
+        },
+        // 生成一个图片验证码的编号，并设置页面中图片验证码img标签的src属性
+        generate_image_code: function(){
+            // 生成一个编号
+            // 严格一点的使用uuid保证编号唯一， 不是很严谨的情况下，也可以使用时间戳
+            this.image_code_id = this.generate_uuid();
 
+            // 设置页面中图片验证码img标签的src属性
+            this.image_code_url = this.host + "/image_codes/" + this.image_code_id + "/";
+        },
         check_pwd: function (){
             var len = this.password.length;
             if(len<8||len>20){
@@ -81,7 +103,14 @@ var vm = new Vue({
                 this.error_phone = true;
             }
         },
-
+        check_image_code: function (){
+            if(!this.image_code) {
+                this.error_image_code_message = '请填写图片验证码';
+                this.error_image_code = true;
+            } else {
+                this.error_image_code = false;
+            }
+        },
         check_sms_code: function(){
             if(!this.sms_code){
                 this.error_sms_code_message = '请填写短信验证码';
@@ -99,14 +128,15 @@ var vm = new Vue({
 
             // 校验参数，保证输入框有数据填写
             this.check_phone();
+            this.check_image_code();
 
-            if (this.error_phone == true) {
+            if (this.error_phone == true || this.error_image_code == true) {
                 this.sending_flag = false;
                 return;
             }
 
             // 向后端接口发送请求，让后端发送短信验证码
-            axios.get(this.host + '/sms_codes/' + this.mobile + '/', {
+            axios.get(this.host + '/sms_codes/' + this.mobile + '/?text=' + this.image_code+'&image_code_id='+ this.image_code_id, {
                     responseType: 'json'
                 })
                 .then(response => {
@@ -131,8 +161,8 @@ var vm = new Vue({
                 })
                 .catch(error => {
                     if (error.response.status == 400) {
-                        this.error_sms_code = error.response.data.message;
-                        this.error_sms_code_message = true;
+                        this.error_image_code_message = '图片验证码有误';
+                        this.error_image_code = true;
                     } else {
                         console.log(error.response.data);
                     }
@@ -145,34 +175,33 @@ var vm = new Vue({
             this.check_phone();
             this.check_sms_code();
 
-            // 访问保存登录用户绑定数据的API
-            axios.post(this.host + '/oauth/qq/user/', {
-                password: this.password,
-                mobile: this.mobile,
-                sms_code: this.sms_code,
-                access_token: this.access_token // 加密openid
-            }, {
-                responseType: 'json',
-                // 让浏览器进行跨域请求时携带cookie
-                withCredentials: true
-            })
-                .then(response => {
-                    // 记录用户登录状态
-                    sessionStorage.clear();
-                    localStorage.clear();
-                    localStorage.token = response.data.token;
-                    localStorage.user_id = response.data.id;
-                    localStorage.username = response.data.username;
-                    location.href = this.get_query_string('state');
+            if (this.error_password == false && this.error_phone == false && this.error_sms_code == false) {
+                axios.post(this.host + '/oauth/sina/user/', {
+                    password: this.password,
+                    mobile: this.mobile,
+                    sms_code: this.sms_code,
+                    access_token: this.access_token
+                }, {
+                    responseType: 'json',
                 })
-                .catch(error => {
-                    if (error.response.status == 400) {
-                        this.error_sms_code_message = error.response.data.message;
-                        this.error_sms_code = true;
-                    } else {
-                        console.log(error.response.data);
-                    }
-                })
+                    .then(response => {
+                        // 记录用户登录状态
+                        sessionStorage.clear();
+                        localStorage.clear();
+                        localStorage.token = response.data.token;
+                        localStorage.user_id = response.data.user_id;
+                        localStorage.username = response.data.username;
+                        location.href = '/';
+                    })
+                    .catch(error => {
+                        if (error.response.status == 400) {
+                            this.error_sms_code_message = error.response.data.message;
+                            this.error_sms_code = true;
+                        } else {
+                            console.log(error.response.data);
+                        }
+                    })
+            }
         }
     }
 });
